restructure explore templates to have header and content and a custom hook for view data just like stories
