+++ assets/src/dashboard/utils/index.js
@@ -19,9 +19,11 @@ export { default as getCurrentYAxis } from './getCurrentYAxis';
 export { default as getFormattedDisplayDate } from './getFormattedDisplayDate';
 export { default as groupBy } from './groupBy';
 export { default as keyboardOnlyOutline } from './keyboardOnlyOutline';
+export { default as throttleToAnimationFrame } from './throttleToAnimationFrame';
 export { default as useFocusOut } from './useFocusOut';
 export { default as useDashboardResultsLabel } from './useDashboardResultsLabel';
 export { default as usePagePreviewSize } from './usePagePreviewSize';
 export { default as useStoryView } from './useStoryView';
-export { default as throttleToAnimationFrame } from './throttleToAnimationFrame';
+export { default as useTemplateView } from './useTemplateView';
+
 export { default as useResizeEffect } from '../../edit-story/utils/useResizeEffect';
